RealTypeConverters: move internals to separate class & add tests
The complicated code, that implements many different converters
and decides which one to use, is now nicely separated from the public
API.

The classes RealTypeConverters.DoubleType previously were public
to allow the use of ClassCopyProvider. No they are package-private
but ClassCopyProcider can still be applied.
